Rename module `input` to avoid naming conflict with element of the same name
Not sure if this was supposed to be allowed, but it was causing https://github.com/rust-lang/rust/issues/96934
